UMT-Hiring-Assistant/requirements.txt
langchain==0.1.0
langchain-community==0.0.13
streamlit==1.29.0
openai==1.5.0
tiktoken==0.5.2
python-dotenv==1.0.0
unstructured==0.11.5
pinecone-client==2.2.4
pypdf==3.17.2
sentence-transformers==2.2.2